CHANGE: Fix PoF test.

+++ tests/pof_tests/pof_test.py
@@ -18,7 +18,8 @@ def x(opt):
     r = params['r']
     theta = params['theta']
 
-    if (constraint_1(opt) < 0) or (constraint_2(opt) < 0):
+    # if (constraint_1(opt) < 0) or (constraint_2(opt) < 0):
+    if constraint_2(opt) < 0:
         raise SolveError
 
     return r * cos(theta)
@@ -53,12 +54,22 @@ def test_pof_basic():
     n_startup_trials = 5
     hv_reference = [0, 0.5]
 
-    path = Path(__file__).parent / 'pof_test.csv'
-    if os.path.exists(path):
-        os.remove(path)
-    path2 = Path(__file__).parent / 'pof_test.db'
-    if os.path.exists(path2):
-        os.remove(path2)
+    # reference が存在しないならば記録モード
+    record_mode = not (Path(__file__).parent / 'pof_reference.csv').is_file()
+
+    # 記録モードならば記録ヒストリを作る
+    if record_mode:
+        path = Path(__file__).parent / 'pof_reference.csv'
+
+    # そうでなければテストヒストリを作る
+    else:
+        path = Path(__file__).parent / 'pof_test.csv'
+        if os.path.exists(path):
+            os.remove(path)
+        path2 = Path(__file__).parent / 'pof_test.db'
+        if os.path.exists(path2):
+            os.remove(path2)
+
 
     # ===== configuration =====
     Sampler = PoFBoTorchSampler
@@ -87,6 +98,8 @@ def test_pof_basic():
     femopt.add_objective(x, direction='maximize', args=femopt.opt)
     femopt.add_objective(y, direction='maximize', args=femopt.opt)
 
+    femopt.add_constraint(constraint_1, lower_bound=0, args=(femopt.opt,))
+
     femopt.set_random_seed(seed)
     femopt._hv_reference = hv_reference
 
@@ -95,25 +108,26 @@ def test_pof_basic():
         confirm_before_exit=False,
     )
 
-    # ===== check =====
-    import numpy as np
-    import pandas as pd
-    from pyfemtet._message.messages import encoding
+    if not record_mode:
+        # ===== check =====
+        import numpy as np
+        import pandas as pd
+        from pyfemtet._message.messages import encoding
 
-    def simplify_df(path):
-        df_ = pd.read_csv(path, encoding=encoding, header=2)
-        pdf_ = pd.DataFrame()
-        pdf_['r'] = df_['r']
-        pdf_['theta'] = df_['theta']
-        pdf_['obj_0'] = df_['obj_0']
-        pdf_['obj_1'] = df_['obj_1']
-        return pdf_.values
+        def simplify_df(path_):
+            df_ = pd.read_csv(path_, encoding=encoding, header=2)
+            pdf_ = pd.DataFrame()
+            pdf_['r'] = df_['r']
+            pdf_['theta'] = df_['theta']
+            pdf_['obj_0'] = df_['obj_0']
+            pdf_['obj_1'] = df_['obj_1']
+            return pdf_.values
 
-    ref = simplify_df(Path(__file__).parent / 'pof_reference.csv')
-    dif = simplify_df(Path(__file__).parent / 'pof_test.csv')
+        ref = simplify_df(Path(__file__).parent / 'pof_reference.csv')
+        dif = simplify_df(Path(__file__).parent / 'pof_test.csv')
 
-    assert (np.abs(dif - ref) / ref).max() < 0.1, "前回結果から 10% 以上の乖離あり"
-    print('Pof test Passed!')
+        assert (np.abs(dif - ref) / ref).max() < 0.1, "前回結果から 10% 以上の乖離あり"
+        print('PoF test Passed!')
 
 
 if __name__ == '__main__':
